refactor: new name for spreadsheet service
